weather_analysis/requirements.txt
celery==5.4.0
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-celery-beat==2.7.0
python-decouple==3.8
requests==2.32.3
redis==5.0.8
